Доработка 3 задания + 4 задание
